python_project/requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
Django==3.2
django-cors-headers==3.7.0
django-environ==0.8.1
django-mysql==3.12.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
environ==1.0
mysqlclient==2.2.6
pillow==11.0.0
PyJWT==2.10.1
pytz==2024.2
setuptools==75.6.0
sqlparse==0.5.3
tzdata==2024.2
